refactor(context): 컨텍스트 관리 클래스 위치 변경
- 컨텍스트 관리 관련 클래스의 위치를 config에서 config.context로 변경
