+++ src/main/java/racingcar/config/context/annotation/Component.java
@@ -1,10 +1,9 @@
-package racingcar.config.annotation;
+package racingcar.config.context.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Component
-public @interface Service {
+public @interface Component {
 
 }
