+++ src/main/java/racingcar/config/context/annotation/Handler.java
@@ -1,4 +1,4 @@
-package racingcar.config.annotation;
+package racingcar.config.context.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
